Many backend changes:
began transition to a new way of message processing. New commands will be processed in a separate object that will hold complete and incomplete messages and commands for already completed messages. After this another object can take and work on those given collections to begin checking for errors etc. Those collections are thread safe, so can be worked on by object that injects commands and object that processes messages further.
Cleaning is to be done, as is writing mentioned further processing class.